Added theme overrides
updated button styling.
renamed ForumModal
