+++ client/src/themes/theme.js
@@ -40,4 +40,30 @@ export const theme = createMuiTheme({
             gradient: 'linear-gradient(45deg, #2574FF, #4B00FF)',
         },
     },
+    overrides: {
+        MuiButton: {
+            textPrimary: {
+                color: '#FFF',
+                background: 'linear-gradient(45deg, #2574FF, #4B00FF)',
+                textTransform: 'none',
+                fontSize: 14,
+                padding: 20,
+                height: '3rem',
+            },
+            outlinedPrimary: {
+                color: '#FFF',
+                textTransform: 'none',
+                fontSize: 14,
+                padding: 20,
+                height: '3rem',
+            },
+            outlinedSecondary: {
+                color: '#2968FF',
+                textTransform: 'none',
+                fontSize: 14,
+                padding: 20,
+                height: '3rem',
+            },
+        },
+    },
 });
